ENH: Basic scalar plotting is functional with pandas
