+++ databroker/muggler/data.py
@@ -549,6 +549,7 @@ class DataMuggler(object):
             return self[attr]
         else:
             raise AttributeError("DataMuggler has no attribute {0} and no "
+
                                   "data source named '{0}'".format(attr))
 
     @property
