+++ databroker/replay/scalar/__init__.py
@@ -3,4 +3,4 @@ import enaml
 with enaml.imports():
     from .view import PlotView, PlotControls
 
-from .model import ScalarModel
+from .model import ScalarCollection
